pyproject.toml
[project]
name = "sfa"
version = "1.0.0"
description = "Official repository for Spatiotemporal Face Alignment for Generalizable Deepfake Detection (FG 2025)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "albumentations>=1.4.1",
    "numpy>=1.26.4",
    "opencv-python>=4.9.0.80",
    "pillow>=10.2.0",
    "scikit-learn>=1.4.1.post1",
    "torch>=2.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sfa"]
